add D_DoomPrefDir() instead of D_DoomExeDir() to iwad_dirs
D_DoomPrefDir() returns the executable directory on Windows,
and a user-writable config directory everywhere else.
