+++ src/d_main.h
@@ -30,7 +30,6 @@ void D_AddFile(const char *file);
 char *D_DoomExeName(void);      // killough 10/98: executable's name
 extern char *basesavegame;     // killough 2/16/98: savegame path
 extern char *screenshotdir; // [FG] screenshot path
-char *D_DoomPrefDir(void); // [FG] default configuration dir
 void D_SetSavegameDirectory(void);
 
 extern const char *gamedescription;
