+++ src/r_data.c
@@ -25,7 +25,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "d_main.h"
+#include "d_iwad.h"
 #include "d_think.h"
 #include "doomdef.h"
 #include "doomstat.h"
